pyproject.toml
[project]
name = "bucso2-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "matplotlib>=3.10.7",
    "numba>=0.62.1",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pytest>=9.0.1",
    "pyyaml>=6.0.3",
    "scipy>=1.16.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]